Add jQuery link, Add JS script.
Added jQuery link in the `head` this allows you to use jQuery within the JS file that will be added. jQuery is qiucker and easier to use than vanilla JS.
A `script` has been added at the bottom of the page which is like the `stylesheet` link in the head. 
Both of these new lines will have to be added to all of the `html` pages so that the `jQuery` will work on the menu.